requirements.txt
# local package
-e .

# external requirements
click~=8.1.3
Sphinx
coverage
flake8
python-dotenv>=0.5.1
pdfminer.six

tqdm~=4.65.0
src~=0.1.0
setuptools~=67.6.1
matplotlib~=3.7.1